pyproject.toml
[project]
name = "paubox-cloud-functions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "cloudevents>=1.11.0",
    "db-dtypes>=1.3.1",
    "functions-framework>=3.8.2",
    "google-api-core>=2.24.0",
    "google-cloud-bigquery>=3.27.0",
    "google-cloud-core>=2.4.1",
    "pandas>=2.2.3",
    "requests>=2.32.3",
    "structlog>=24.4.0",
]

[tool.uv]
dev-dependencies = [
    "black>=24.10.0",
    "mypy>=1.13.0",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
]

[tool.ruff]
line-length = 88
